Add FUSE pulldown and PTC dummy
In contrast to documentation provided by TI FUSE is neither just an
output nor does the PTCEN feature work as advertised.

The FUSE pin contains a switchable 150nA current source. Presumably it
is used as a safety to detect if the FUSE pin is open. Unfortunately the
gauage will not allow control of the CHG, DSG and PRECHG FETs when it
detects an open FUSE pin. There does not seem to be any configuration
option for this behaviour either.

The PTC pin allows connecting a PTC to detect overheating of the main
power path control FETs. According to the BQ40Z50 datasheet it is possible
to disable this safety function by tieing PTCEN to VSS. Unfortunately this
does not seem to work as intended. When tieing both PTCEN and PTC to VSS
the gauage does not give any indication of a failure but refuses to switch
any of CHG, PRECHG or DSG FETs.